# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GameForge-Copilot",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/me/usage": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "My Usage",
        "operationId": "my_usage_api_v1_me_usage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UsageResp_"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/usage": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "Admin Usage",
        "operationId": "admin_usage_api_v1_admin_usage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AdminUsageResp_"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/me/usage/breakdown": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "Usage Breakdown",
        "operationId": "usage_breakdown_api_v1_me_usage_breakdown_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(game|run)$",
              "default": "game",
              "title": "Scope"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedData_UsageBreakdownItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/games/{game_id}/usage": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "Game Usage",
        "operationId": "game_usage_api_v1_games__game_id__usage_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_GameUsageResp_"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/games/{game_id}/analytics": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "Game Analytics",
        "operationId": "game_analytics_api_v1_games__game_id__analytics_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_GameAnalyticsResp_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/analytics/top": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "Admin Analytics Top",
        "operationId": "admin_analytics_top_api_v1_admin_analytics_top_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_dict__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminUsageResp": {
        "properties": {
          "system": {
            "$ref": "#/components/schemas/SystemUsage"
          },
          "top_users": {
            "items": {
              "$ref": "#/components/schemas/AdminUserUsage"
            },
            "type": "array",
            "title": "Top Users"
          }
        },
        "type": "object",
        "required": [
          "system",
          "top_users"
        ],
        "title": "AdminUsageResp"
      },
...
      "AdminUserUsage": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "month_input_tokens": {
            "type": "integer",
            "title": "Month Input Tokens"
          },
          "month_output_tokens": {
            "type": "integer",
            "title": "Month Output Tokens"
          },
          "calls": {
            "type": "integer",
            "title": "Calls"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "email",
          "month_input_tokens",
          "month_output_tokens",
          "calls"
        ],
        "title": "AdminUserUsage"
      },
...
      "ApiResponse_AdminUsageResp_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminUsageResp"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[AdminUsageResp]"
      },
...
      "ApiResponse_GameAnalyticsResp_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GameAnalyticsResp"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[GameAnalyticsResp]"
      },
...
      "ApiResponse_GameUsageResp_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GameUsageResp"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[GameUsageResp]"
      },
...
      "ApiResponse_UsageResp_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UsageResp"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[UsageResp]"
      },
...
      "ApiResponse_list_dict__": {
        "properties": {
          "data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[list[dict]]"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "统一错误响应：{\"error\": {code, message, detail}}"
      },
      "GameAnalyticsResp": {
        "properties": {
          "game_id": {
            "type": "string",
            "format": "uuid",
            "title": "Game Id"
          },
          "play_count": {
            "type": "integer",
            "title": "Play Count"
          },
          "pv_30d": {
            "type": "integer",
            "title": "Pv 30D"
          },
          "uv_30d": {
            "type": "integer",
            "title": "Uv 30D"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "play_count",
          "pv_30d",
          "uv_30d"
        ],
        "title": "GameAnalyticsResp"
      },
...
      "GameUsageResp": {
        "properties": {
          "game_id": {
            "type": "string",
            "format": "uuid",
            "title": "Game Id"
          },
          "month": {
            "$ref": "#/components/schemas/UsageBucket"
          },
          "estimated_usd": {
            "type": "number",
            "title": "Estimated Usd"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "month",
          "estimated_usd"
        ],
        "title": "GameUsageResp"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedData_UsageBreakdownItem_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/UsageBreakdownItem"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "size"
        ],
        "title": "PaginatedData[UsageBreakdownItem]"
      },
...
      "QuotaInfo": {
        "properties": {
          "daily_token_limit": {
            "type": "integer",
            "title": "Daily Token Limit"
          },
          "daily_used": {
            "type": "integer",
            "title": "Daily Used"
          },
          "remaining": {
            "type": "integer",
            "title": "Remaining"
          }
        },
        "type": "object",
        "required": [
          "daily_token_limit",
          "daily_used",
          "remaining"
        ],
        "title": "QuotaInfo"
      },
...
      "SystemUsage": {
        "properties": {
          "today": {
            "$ref": "#/components/schemas/UsageBucket"
          },
          "month": {
            "$ref": "#/components/schemas/UsageBucket"
          },
          "total": {
            "$ref": "#/components/schemas/UsageBucket"
          }
        },
        "type": "object",
        "required": [
          "today",
          "month",
          "total"
        ],
        "title": "SystemUsage"
      },
...
      "UsageBreakdownItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens"
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens"
          },
          "calls": {
            "type": "integer",
            "title": "Calls"
          },
          "estimated_usd": {
            "type": "number",
            "title": "Estimated Usd"
          }
        },
        "type": "object",
        "required": [
          "id",
          "input_tokens",
          "output_tokens",
          "calls",
          "estimated_usd"
        ],
        "title": "UsageBreakdownItem"
      },
      "UsageBucket": {
        "properties": {
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens"
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens"
          },
          "calls": {
            "type": "integer",
            "title": "Calls"
          }
        },
        "type": "object",
        "required": [
          "input_tokens",
          "output_tokens",
          "calls"
        ],
        "title": "UsageBucket"
      },
      "UsageResp": {
        "properties": {
          "today": {
            "$ref": "#/components/schemas/UsageBucket"
          },
          "month": {
            "$ref": "#/components/schemas/UsageBucket"
          },
          "total": {
            "$ref": "#/components/schemas/UsageBucket"
          },
          "quota": {
            "$ref": "#/components/schemas/QuotaInfo"
          }
        },
        "type": "object",
        "required": [
          "today",
          "month",
          "total",
          "quota"
        ],
        "title": "UsageResp"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}